@@ -1,6 +1,6 @@
 module github.com/d-kuro/gwq
 
-go 1.24.2
+go 1.26.0
 
 require (
 	github.com/bmatcuk/doublestar/v4 v4.10.0
